"Updated Personnel API code: refactored authentication and permission middlewares, modified controllers and routes for departments, personnel, and tokens, and made minor formatting changes."

+++ class-notes/20-PersonelApi_Token_Perms/src/controllers/auth.js
@@ -1,10 +1,10 @@
-"use strict";
+"use strict"
 /* -------------------------------------------------------
     EXPRESS - Personnel API
 ------------------------------------------------------- */
-const Personnel = require("../models/personnel");
-const passwordEncrypt = require("../helpers/passwordEncrypt");
-const Token = require("../models/token");
+const Personnel = require("../models/personnel")
+const passwordEncrypt = require('../helpers/passwordEncrypt')
+const Token = require('../models/token')
 
 /* ------------------------------------------------------- *
 // Seesion 
@@ -69,55 +69,66 @@ module.exports = {
 
 /* ------------------------------------------------------- */
 
+
 module.exports = {
-  login: async (req, res) => {
-    const { username, password } = req.body;
 
-    if (username && password) {
-      const user = await Personnel.findOne({ username, password });
+    login: async (req, res) => {
 
-      if (user) {
-        if (user.isActive) {
-          /* TOKEN*/
-          let tokenData = await Token.findOne({ userId: user._id });
-          if (!tokenData) {
-            tokenData = Token.create({
-              userId: user._id,
-              token: user._id + Date.now(),
-            });
-          }
+        const { username, password } = req.body
 
-          res.status(201).send({
-            error: false,
-            token:tokenData.token,
-            user,
-          });
+        if (username && password) {
+
+            const user = await Personnel.findOne({ username, password })
+
+            if (user) {
+
+                if (user.isActive) {
+
+                    /* TOKEN */
+
+                    let tokenData = await Token.findOne({ userId: user._id })
+
+                    if (!tokenData) {
+                        tokenData = Token.create({
+                            userId: user._id,
+                            token: passwordEncrypt(user._id + Date.now())
+                        })
+                    }
+
+
+                    res.status(201).send({
+                        error: false,
+                        token: tokenData.token,
+                        user
+                    })
+
+                } else {
+                    res.errorStatusCode = 401
+                    throw new Error("This user is not active.")
+                }
+
+
+
+            } else {
+                res.errorStatusCode = 401
+                throw new Error("Wrong username or password.")
+            }
+
         } else {
-          res.errorStatusCode = 401;
-          throw new Error("This user is not active");
+            res.errorStatusCode = 401
+            throw new Error("Please enter username and password.")
         }
-      } else {
-        res.errorStatusCode = 401;
-        throw new Error("Wrong username or password");
-      }
-    } else {
-      res.errorStatusCode = 401;
-      throw new Error("Please enter username and password");
-    }
-  },
 
-  logout: async (req, res) => {
+    },
 
+    logout: async (req, res) => {
 
-    const data = req.user ? await Token.deleteOne({
-        userId: req.user._id
-    }) : null
+        const data = req.user ? await Token.deleteOne({ userId: req.user._id }) : null
 
-
-    res.status(200).send({
-        error:false,
-        message:'Logout Success',
-        data
-    })
-  },
-};
+        res.status(200).send({
+            error: false,
+            message: 'Logout Success.',
+            data
+        })
+    },
+}
